update unit tests under abm.test

+++ activitysim/core/interaction_simulate.py
@@ -85,28 +85,30 @@ def eval_interaction_utilities(
     buffer = io.StringIO()
     df.info(memory_usage="deep", buf=buffer, verbose=True, show_counts=True)
     s = buffer.getvalue()
-    with open(
-        os.path.join(
-            state.filesystem.output_dir, trace_label + ".interaction_df.info.txt"
-        ),
-        "w",
-        encoding="utf-8",
-    ) as f:
-        f.write(s)
+    # check if output dir exists
+    if os.path.exists(state.filesystem.output_dir):
+        with open(
+            os.path.join(
+                state.filesystem.output_dir, trace_label + ".interaction_df.info.txt"
+            ),
+            "w",
+            encoding="utf-8",
+        ) as f:
+            f.write(s)
 
-    df.memory_usage(deep=True).to_csv(
-        os.path.join(
-            state.filesystem.output_dir,
-            trace_label + ".interaction_df.memory_usage_deep.txt",
+        df.memory_usage(deep=True).to_csv(
+            os.path.join(
+                state.filesystem.output_dir,
+                trace_label + ".interaction_df.memory_usage_deep.txt",
+            )
         )
-    )
 
-    df.memory_usage().to_csv(
-        os.path.join(
-            state.filesystem.output_dir,
-            trace_label + ".interaction_df.memory_usage.txt",
+        df.memory_usage().to_csv(
+            os.path.join(
+                state.filesystem.output_dir,
+                trace_label + ".interaction_df.memory_usage.txt",
+            )
         )
-    )
 
     process = psutil.Process(os.getpid())
     logger.info(
